chore: Remove commented code and update navigation in PractiseScreen
The code changes in the `PractiseScreen` include removing commented code and updating the navigation by adding a back button with an arrow icon. This improves the user experience by allowing users to easily navigate back to the previous screen. The changes also include updating the title of the app bar to "Practice" and displaying the `PractiseConsumers` widget in the body of the screen.

This commit improves the codebase by removing unnecessary code and enhancing the navigation and user interface of the `PractiseScreen`.